(tests_rand_start): Make shift well-defined.

+++ tests/misc.c
@@ -116,7 +116,7 @@ tests_rand_start (void)
 #if HAVE_GETTIMEOFDAY
           struct timeval  tv;
           gettimeofday (&tv, NULL);
-          seed = tv.tv_sec ^ (tv.tv_usec << 12);
+          seed = tv.tv_sec ^ ((unsigned long) tv.tv_usec << 12);
 	  seed &= 0xffffffff;
 #else
           time_t  tv;
